fix: 서버컴포넌트, 클라이언트컴포넌트 분리
